ct dk

+++ MVC/models/Admin/dang_ky.php
@@ -14,41 +14,10 @@ function layDSDK_1() {
   
 }
 
-function MocKH_0()
-{ 
-  $sql = "SELECT khach_hang.ten_khach_hang,khach_hang.ma_kh
-  FROM khach_hang,don
-  WHERE khach_hang.ma_kh = don.ma_kh  AND trang_thai = 0";
-
-    $kh1 = getData($sql, FETCH_ALL);
-    return $kh1;
-}function MocKH_1()
-{ 
-  $sql = "SELECT khach_hang.ten_khach_hang,khach_hang.ma_kh
-  FROM khach_hang,don
-  WHERE khach_hang.ma_kh = don.ma_kh AND trang_thai = 1";
-
-    $kh1 = getData($sql, FETCH_ALL);
-    return $kh1;
-}
-
-function MocXe_0()
-{
-
-  $sql = "SELECT xe.ten_xe,xe.ma_xe
-  FROM xe,don
-  WHERE xe.ma_xe = don.ma_xe  AND trang_thai = 0";
-    $xe1 = getData($sql, FETCH_ALL);
-    return $xe1;
-}
-function MocXe_1()
-{
-
-  $sql = "SELECT xe.ten_xe,xe.ma_xe
-  FROM xe,don
-  WHERE xe.ma_xe = don.ma_xe  AND trang_thai = 1";
-    $xe1 = getData($sql, FETCH_ALL);
-    return $xe1;
+function ctDangKy($ma_dat){
+  $sql = "SELECT * FROM xe,don,khach_hang WHERE xe.ma_xe = don.ma_xe AND khach_hang.ma_kh = don.ma_kh AND ma_dat = $ma_dat";
+  $ct = getData($sql, FETCH_ONE);
+  return $ct;
 }
 function suaDangKy()
 {  
@@ -56,21 +25,26 @@ function suaDangKy()
    
   $trang_thai = '1';
         $ma_dat = $_GET['ma_dat'];
-        $sql = "UPDATE don SET trang_thai='$trang_thai' WHERE  ma_dat  =$ma_dat";
-        $conn = getConnect();
-        $statement = $conn->prepare($sql);
-        $statement->execute();
-        header('Location:index2.php?url=ds_dang_ky_1');
+        if(isset($_POST['btn_duyet'])){
+          $sql = "UPDATE don SET trang_thai='$trang_thai' WHERE  ma_dat  =$ma_dat";
+          $conn = getConnect();
+          $statement = $conn->prepare($sql);
+          $statement->execute();
+          header('Location:index2.php?url=ds_dang_ky_1');
+        }
+
   
 }
 function xoaDangKy()
 {  
     $ma_dat = $_GET['ma_dat'];   
+    if(isset($_POST['btn_delete'])){
         $sql = "DELETE FROM don WHERE ma_dat=$ma_dat ";
         $conn = getConnect();
         $statement = $conn->prepare($sql);
         $statement->execute();
         header('Location:index2.php?url=ds_dang_ky_1');
+    }
       }
